vmm: rename devices module to serial
The virtio device implementations are going to be added to a local
devices crate, so renaming the module with the same name within vmm
to serial (because there's only a serial device there for the time
being). Going foward, we should move the serial console to the devices
crate as well.
